better comments and tests
