+++ src/common/html-to-text.test.ts
@@ -4,4 +4,7 @@ test('works', () => {
   expect(htmlToText('foo &gt; &amp;gt; <br> <b>foooo</b> aa')).toBe(
     'foo > &gt;  foooo aa',
   );
+  expect(htmlToText('String with <b>html</b> &amp; entities &lt;&gt;')).toBe(
+    'String with html & entities <>',
+  );
 });
